requirements.txt
pydantic>=2
python-dotenv>=1
httpx>=0.27
tenacity>=8
orjson>=3
eth-abi>=5
pydantic-settings>=2
aiohttp>=3.9
cryptography>=41.0
lighthouseweb3>=0.1.5
# uvloop: platform-specific; install manually on Linux/macOS for better async performance
uvloop
